# project/tests/goldens/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Computational Metaphysics Engine",
    "description": "Modular 10-Domain Metaphysical Calculation, AI Debate & Multi-Agent Engine",
    "version": "1.0.0.34833bc"
  },
  "paths": {
...
    "/api/v1/luopan/calculate": {
      "post": {
        "tags": [
          "LuoPan Compass & Dream Interpretation"
        ],
        "summary": "Calculate Luopan",
        "description": "Calculate 24-Mountain compass orientation and Period 9 Flying Star 9-Palace sector heatmap.",
        "operationId": "calculate_luopan_api_v1_luopan_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LuoPanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Calculate Luopan Api V1 Luopan Calculate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dream/interpret": {
      "post": {
        "tags": [
          "LuoPan Compass & Dream Interpretation"
        ],
        "summary": "Interpret Dream",
        "description": "Decode dream symbolism into I Ching 64 Hexagrams and Sattaleka lucky omen numbers.",
        "operationId": "interpret_dream_api_v1_dream_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Interpret Dream Api V1 Dream Interpret Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DreamRequest": {
        "properties": {
          "dream_text": {
            "type": "string",
            "title": "Dream Text",
            "description": "Description of the dream in Thai or English"
          },
          "user_day_master": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Day Master",
            "description": "User's BaZi Day Master Stem"
          }
        },
        "type": "object",
        "required": [
          "dream_text"
        ],
        "title": "DreamRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LuoPanRequest": {
        "properties": {
          "facing_degree": {
            "type": "number",
            "title": "Facing Degree",
            "description": "Facing compass degree (0 - 360)",
            "default": 180.0
          },
          "period": {
            "type": "integer",
            "title": "Period",
            "description": "Feng Shui Period (default 9: 2024-2043)",
            "default": 9
          }
        },
        "type": "object",
        "title": "LuoPanRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}